ElectroPhysiology Methods

using ElectroPhysiology

These methods become available after this command is run.

Stimulus Protocols

ElectroPhysiology.Stimulus — Type

Stimulus

An abstract type representing a stimulus in a physiological experiment.

Subtypes of Stimulus should implement specific stimulus types and their corresponding parameters.

ElectroPhysiology.Flash — Type

Flash <: Stimulus

A Flash is a subtype of Stimulus representing a flash stimulus in a physiological experiment.

Fields

intensity: A Real value indicating the intensity of the flash.

Constructors

Flash(): Creates a default Flash object with an intensity of 0.0.

Flash(intensity::Real): Creates a Flash object with the specified intensity.

ElectroPhysiology.StimulusProtocol — Type

StimulusProtocol{T, S} where {T <: Real, S <: Stimulus}

A mutable struct representing a stimulus protocol for physiological data.

Fields

type: A Stimulus object describing the type of stimulus applied during the experiment.

channelName: A Union{String, Int64} representing the name or number of the channel where the stimulus is applied.

timestamps: A Vector{Tuple{T, T}} storing the start and end timestamps of the stimulus events.

Constructors

StimulusProtocol(): Creates a default StimulusProtocol object with Flash() stimulus, "Nothing" channel, and a single (0.0, 0.0) timestamp.

StimulusProtocol(stimulus_channel::String): Creates a StimulusProtocol object with Flash() stimulus, the provided stimulus_channel, and a single (0.0, 0.0) timestamp.

StimulusProtocol(swp::Int64): Creates a StimulusProtocol object with Flash() stimulus, "Nothing" channel, and swp number of (0.0, 0.0) timestamps.

StimulusProtocol(stimulus_channel::String, swp::Int64): Creates a StimulusProtocol object with Flash() stimulus, the provided stimulus_channel, and swp number of (0.0, 0.0) timestamps.

Examples

stim1 = StimululsProtocol()
stim_channel = StimulusProtocol("IN 3")
stim_sweep = StimulusProtocol(3)
stim_sweep_channel = StimululsProtocol(3, "IN 3")
stim_tstamp = StimululsProtocol((0.0, 0.01))
stim_type_time_channel = StimulusProtocol(Flash(), "IN 3", (0.00, 0.01))

ElectroPhysiology.extractStimulus — Function

extractStimulus(abfInfo::Dict{String, Any}; stimulus_name::String="IN 7", stimulus_threshold::Float64=2.5)
extractStimulus(abf_path::String; kwargs...)

Extract the stimulus information from the given abfInfo dictionary and returns a StimulusProtocol object containing stimulus timestamps.

Arguments

abfInfo: A dictionary containing information about the physiological data.

stimulus_name: (Optional) The name of the stimulus channel. Default is "IN 7".

stimulus_threshold: (Optional) The threshold for detecting stimulus events in the waveform. Default is 2.5.

Returns

A StimulusProtocol object containing the stimulus timestamps for each trial.

Examples

abfInfo = loadABF("path/to/abf/file")
stimuli = extractStimulus(abfInfo)

stimuli = extractStimulus("path/to/abf/file")

ElectroPhysiology.setIntensity — Function

setIntensity(stimulus_protocols::StimulusProtocol{T, Flash}, photons::Vector{T}) where T<:Real
setIntensity(stimulus_protocols::StimulusProtocol{T, Flash}, photon::T) where T<:Real

This allows the intensity of the stimulus protocol (or multiple stimulus protocols be set).

Arguments

stimulus_protocol::StimulusProtocol{T, S} where S <: Flash: A stimulus protocol

photons::Vector: a vector of numbers representing the photon amount.

Examples

ElectroPhysiology.getIntensity — Function

setIntensity(stimulus_protocols::StimulusProtocol{T, Flash}, photons::Vector)

This allows the intensity of the stimulus protocol (or multiple stimulus protocols be set).

Arguments

stimulus_protocol::StimulusProtocol{T, S} where S <: Flash: A stimulus protocol

photons::Vector: a vector of numbers representing the photon amount.

Structs

ElectroPhysiology.Experiment — Type

Experiment{T}

A mutable struct representing a physiological experiment.

Fields

HeaderDict: A dictionary containing header information for the experiment.

dt: A Real value representing the time step between data points.

t: A vector containing the time points of the experiment.

data_array: A 3-dimensional array containing the experimental data.

chNames: A vector of strings representing the names of the channels.

chUnits: A vector of strings representing the units of the channels.

chTelegraph: A vector of Real values representing the telegraph values of the channels.

stimulus_protocol: A StimulusProtocol{T} object containing the stimulus protocol information.

Constructors

Experiment(data_array::AbstractArray; data_idx = 2): Create an Experiment object from an input data array with an optional data index.

Experiment(time::Vector, data_array::Array{T, 3}) where T <: Real: Create an Experiment object from an input time vector and data array.

Functions

ElectroPhysiology.pad — Method

pad(trace::Experiment{T}, n_add::Int64; position::Symbol=:post, val::T=0.0) where {T <: Real}
pad!(trace::Experiment{T}, n_add::Int64; position::Symbol=:post, val::T=0.0) where {T <: Real}

Pad the data in a given Experiment object by adding elements with a specified value val either before or after the existing data. This function returns a new Experiment object with the padded data. pad! is the inplace version, while pad creates a new experiment

Arguments

trace: An Experiment{T} object containing the experimental data.

n_add: The number of elements to add.

position: (Optional) A symbol specifying where to add the elements. Can be :pre (before) or :post (after). Default is :post.

val: (Optional) The value to use for padding. Default is 0.0.

Example

exp = Experiment(data_array)
padded_exp = pad(exp, 100, position=:pre, val=0.0)

Experiment readers

ElectroPhysiology.readABF — Method

readABF(::Type{T}, abf_data::Union{String,Vector{UInt8}};
    trials::Union{Int64,Vector{Int64}}=-1,
    channels::Union{Int64, Vector{String}}=["Vm_prime", "Vm_prime4"],
    average_trials::Bool=false,
    stimulus_name::Union{String, Vector{String}, Nothing}="IN 7",
    stimulus_threshold::T=2.5,
    warn_bad_channel=false,
    flatten_episodic::Bool=false,
    time_unit=:s,
) where {T<:Real}

Read an Axon Binary File (ABF) and return an Experiment object. The function extracts data for the specified trials and channels, and optionally averages trials or flattens episodic data.

Arguments

abf_data: A String representing the ABF file path or a Vector{UInt8} containing the ABF file content.

trials: An Int64 or a Vector{Int64} specifying the trials to extract. Default is -1 (all trials).

channels: An Int64 or a Vector{String} specifying the channels to extract. Default is ["Vmprime", "Vmprime4"].

average_trials: A Bool specifying whether to average the trials. Default is false.

stimulus_name: A String, Vector{String}, or Nothing specifying the stimulus name(s). Default is "IN 7".

stimulus_threshold: A threshold value of type T for the stimulus. Default is 2.5.

warn_bad_channel: A Bool specifying whether to warn if a channel is improper. Default is false.

flatten_episodic: A Bool specifying whether to flatten episodic stimulation to be continuous. Default is false.

time_unit: A Symbol specifying the time unit. Default is :s (seconds).

Returns

An Experiment object containing the extracted data, along with metadata.

Example

exp = readABF(Float32, "path/to/abf_file.abf")