Methods

ElectroPhysiology Methods

using ElectroPhysiology

These methods become available after this command is run.

Stimulus Protocols

ElectroPhysiology.StimulusProtocol — Type

StimulusProtocol{T, S}

A mutable struct representing a stimulus protocol for physiological data.

Fields

type: A Stimulus object describing the type of stimulus applied during the experiment.

channelName: A Union{String, Int64} representing the name or number of the channel where the stimulus is applied.

timestamps: A Vector{Tuple{T, T}} storing the start and end timestamps of the stimulus events.

Constructors

StimulusProtocol(): Creates a default StimulusProtocol object with Flash() stimulus, "Nothing" channel, and a single (0.0, 0.0) timestamp.

StimulusProtocol(stimulus_channel::String): Creates a StimulusProtocol object with Flash() stimulus, the provided stimulus_channel, and a single (0.0, 0.0) timestamp.

StimulusProtocol(swp::Int64): Creates a StimulusProtocol object with Flash() stimulus, "Nothing" channel, and swp number of (0.0, 0.0) timestamps.

StimulusProtocol(stimulus_channel::String, swp::Int64): Creates a StimulusProtocol object with Flash() stimulus, the provided stimulus_channel, and swp number of (0.0, 0.0) timestamps.

ElectroPhysiology.extractStimulus — Function

extractStimulus(abfInfo::Dict{String, Any}; stimulus_name::String="IN 7", stimulus_threshold::Float64=2.5)
extractStimulus(abf_path::String; kwargs...)

Extract the stimulus information from the given abfInfo dictionary and returns a StimulusProtocol object containing stimulus timestamps.

Arguments

abfInfo: A dictionary containing information about the physiological data.

stimulus_name: (Optional) The name of the stimulus channel. Default is "IN 7".

stimulus_threshold: (Optional) The threshold for detecting stimulus events in the waveform. Default is 2.5.

Returns

A StimulusProtocol object containing the stimulus timestamps for each trial.

Examples

abfInfo = loadABF("path/to/abf/file")
stimuli = extractStimulus(abfInfo)

stimuli = extractStimulus("path/to/abf/file")

Missing docstring.

Missing docstring for setIntensity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getIntensity. Check Documenter's build log for details.

Structs

ElectroPhysiology.Experiment — Type

Experiment{T}

A mutable struct representing a physiological experiment.

Fields

HeaderDict: A dictionary containing header information for the experiment.

dt: A Real value representing the time step between data points.

t: A vector containing the time points of the experiment.

data_array: A 3-dimensional array containing the experimental data.

chNames: A vector of strings representing the names of the channels.

chUnits: A vector of strings representing the units of the channels.

chTelegraph: A vector of Real values representing the telegraph values of the channels.

stimulus_protocol: A StimulusProtocol{T} object containing the stimulus protocol information.

Constructors

Experiment(data_array::AbstractArray; data_idx = 2): Create an Experiment object from an input data array with an optional data index.

Experiment(time::Vector, data_array::Array{T, 3}) where T <: Real: Create an Experiment object from an input time vector and data array.

Functions

ElectroPhysiology.pad — Method

pad(trace::Experiment{T}, n_add::Int64; position::Symbol=:post, val::T=0.0) where {T <: Real}
pad!(trace::Experiment{T}, n_add::Int64; position::Symbol=:post, val::T=0.0) where {T <: Real}

Pad the data in a given Experiment object by adding elements with a specified value val either before or after the existing data. This function returns a new Experiment object with the padded data. pad! is the inplace version, while pad creates a new experiment

Arguments

trace: An Experiment{T} object containing the experimental data.

n_add: The number of elements to add.

position: (Optional) A symbol specifying where to add the elements. Can be :pre (before) or :post (after). Default is :post.

val: (Optional) The value to use for padding. Default is 0.0.

Example

exp = Experiment(data_array)
padded_exp = pad(exp, 100, position=:pre, val=0.0)

PhysiologyAnalysis Methods

These methods only become available after

using PhysiologyAnalysis

is run.

PhysiologyPlotting Methods

These methods only become available after

using PhysiologyPlotting

is run.